Extracted testing of create binding job to shared example
Shared example can be used to test route and credential type of bindings.
Next step in the refactoring is merging the action in one single version.
