+++ app/models/services/service_binding.rb
@@ -144,6 +144,11 @@ module VCAP::CloudController
       false
     end
 
+    def save_with_attributes_and_new_operation( attributes, operation)
+      save_with_new_operation(operation, attributes: attributes)
+      self
+    end
+
     def save_with_new_operation(last_operation, attributes: {})
       ServiceBinding.db.transaction do
         self.lock!
