+++ spec/unit/jobs/v3/create_service_binding_job_factory_spec.rb
@@ -1,3 +1,4 @@
+require 'db_spec_helper'
 require 'jobs/v3/create_service_binding_job_factory'
 
 module VCAP::CloudController
@@ -8,14 +9,35 @@ module VCAP::CloudController
       describe '#for' do
         it 'should return route job actor when type is route' do
           actor = CreateServiceBindingFactory.for(:route)
-          expect(actor).to be_an_instance_of(CreateRouteBindingJobActor)
+          expect(actor).to be_an_instance_of(CreateServiceRouteBindingJobActor)
         end
 
-        it 'should return route job actor when type is credential' do
+        it 'should return credential job actor when type is credential' do
           actor = CreateServiceBindingFactory.for(:credential)
           expect(actor).to be_an_instance_of(CreateServiceCredentialBindingJobActor)
         end
+
+        it 'raise for unknown types' do
+          expect { CreateServiceBindingFactory.for(:key) }.to raise_error(CreateServiceBindingFactory::InvalidType)
+        end
       end
+
+      describe '#action' do
+        it 'should return route action when type is route' do
+          actor = CreateServiceBindingFactory.action(:route, {}, {})
+          expect(actor).to be_an_instance_of(ServiceRouteBindingCreate)
+        end
+
+        it 'should return credential binding action when type is credential' do
+          actor = CreateServiceBindingFactory.action(:credential, {}, {})
+          expect(actor).to be_an_instance_of(ServiceCredentialBindingCreate)
+        end
+
+        it 'raise for unknown types' do
+          expect { CreateServiceBindingFactory.action(:key, {}, {}) }.to raise_error(CreateServiceBindingFactory::InvalidType)
+        end
+      end
+
     end
   end
 end
